fix: group control for search_ill & unescape
escaped `&`, `[` and `]` handled

+++ src/handlers/search_ill/search_ill.ts
@@ -15,8 +15,10 @@ import {
 
 import {
   cq_image,
+  is_at_self,
   remove_cqcode,
   send_group_at_message,
+  unescape_non_cq,
 } from "../../cqhttp.ts";
 
 class Context {
@@ -159,48 +161,50 @@ function check_rate_limit(group_id: number) {
 }
 
 export async function search_ill_handler(report: Report) {
-  const input = remove_cqcode(report.message).trim();
-  if (config.ill_input_match.some((m) => input.includes(m))) {
-    log("search_ill");
+  if (is_at_self(report.message) && groups.includes(report.group_id)) {
+    const input = unescape_non_cq(remove_cqcode(report.message).trim());
+    if (config.ill_input_match.some((m) => input.includes(m))) {
+      log("search_ill");
 
-    if (!check_rate_limit(report.group_id)) {
-      send_group_at_message(
-        report.group_id,
-        `Rate limit exceeded, current rate limit is ${config.rate_limit_per_hour} per hour`,
-        report.sender.user_id,
-      );
-      return;
-    }
-
-    const ids = await call_function(await get_tag_reply(input), report);
-    const pairs = await download_small(ids);
-    if (pairs.length == 0) {
-      log("No ill found");
-      send_group_at_message(
-        report.group_id,
-        "再怎么找也找不到啦>_<",
-        report.sender.user_id,
-      );
-      return;
-    }
-    log(pairs);
-
-    const choice = await choose_ill(pairs.map(([_, path]) => path));
-    let url = "";
-    pairs.forEach(([id, path], index) => {
-      Deno.removeSync(path);
-      if (index == choice) {
-        url = `https://www.pixiv.net/artworks/${id}`;
-        log("choice: ", url);
+      if (!check_rate_limit(report.group_id)) {
+        send_group_at_message(
+          report.group_id,
+          `Rate limit exceeded, current rate limit is ${config.rate_limit_per_hour} per hour`,
+          report.sender.user_id,
+        );
+        return;
       }
-    });
 
-    const [[_, path]] = await download_large([ids[choice]]);
-    send_group_at_message(
-      report.group_id,
-      [cq_image(Deno.readFileSync(path)), url].join("\n"),
-      report.sender.user_id,
-    );
+      const ids = await call_function(await get_tag_reply(input), report);
+      const pairs = await download_small(ids);
+      if (pairs.length == 0) {
+        log("No ill found");
+        send_group_at_message(
+          report.group_id,
+          "再怎么找也找不到啦>_<",
+          report.sender.user_id,
+        );
+        return;
+      }
+      log(pairs);
+
+      const choice = await choose_ill(pairs.map(([_, path]) => path));
+      let url = "";
+      pairs.forEach(([id, path], index) => {
+        Deno.removeSync(path);
+        if (index == choice) {
+          url = `https://www.pixiv.net/artworks/${id}`;
+          log("choice: ", url);
+        }
+      });
+
+      const [[_, path]] = await download_large([ids[choice]]);
+      send_group_at_message(
+        report.group_id,
+        [cq_image(Deno.readFileSync(path)), url].join("\n"),
+        report.sender.user_id,
+      );
+    }
   }
 }
